Move playlist policy namespace weekly -> generated
