+++ src/Jellyfin.Plugin.ListenBrainz/Tasks/SyncGeneratedPlaylists/PlaylistCandidate.cs
@@ -1,5 +1,5 @@
 using Jellyfin.Plugin.ListenBrainz.Api.Models;
 
-namespace Jellyfin.Plugin.ListenBrainz.Tasks.SyncWeeklyPlaylists;
+namespace Jellyfin.Plugin.ListenBrainz.Tasks.SyncGeneratedPlaylists;
 
 internal sealed record PlaylistCandidate(Playlist Playlist, PlaylistType Type);
