+++ src/Jellyfin.Plugin.ListenBrainz/Tasks/SyncGeneratedPlaylists/PlaylistTypeDescriptor.cs
@@ -1,6 +1,6 @@
 using Jellyfin.Plugin.ListenBrainz.Configuration;
 
-namespace Jellyfin.Plugin.ListenBrainz.Tasks.SyncWeeklyPlaylists;
+namespace Jellyfin.Plugin.ListenBrainz.Tasks.SyncGeneratedPlaylists;
 
 /// <summary>
 /// Describes a <see cref="PlaylistType"/>: how to recognize it, how it is retained,
