+++ tests/Jellyfin.Plugin.ListenBrainz.Tests/Tasks/WeeklyPlaylists/WeeklyRotationTests.cs
@@ -5,7 +5,7 @@ using Jellyfin.Plugin.ListenBrainz.Api.Models;
 using Jellyfin.Plugin.ListenBrainz.Configuration;
 using Jellyfin.Plugin.ListenBrainz.Dtos;
 using Jellyfin.Plugin.ListenBrainz.Tasks;
-using Jellyfin.Plugin.ListenBrainz.Tasks.SyncWeeklyPlaylists;
+using Jellyfin.Plugin.ListenBrainz.Tasks.SyncGeneratedPlaylists;
 using Xunit;
 
 namespace Jellyfin.Plugin.ListenBrainz.Tests.Tasks.WeeklyPlaylists;
